move BWS route modules into routes folder
