+++ packages/bitcore-wallet-service/src/lib/expressapp.ts
@@ -3,22 +3,22 @@ import express from 'express';
 import 'source-map-support/register';
 import config from '../config';
 import * as Types from '../types/expressapp';
-import { registerAdvertisementRoutes } from './expressapp/advertisements';
-import { registerBlockchainRoutes } from './expressapp/blockchain';
-import { createRouteHelpers } from './expressapp/context';
-import { registerMaintenanceRoutes } from './expressapp/maintenance';
-import { registerMoralisRoutes } from './expressapp/moralis';
-import { registerNotificationRoutes } from './expressapp/notifications';
-import { registerServiceRoutes } from './expressapp/services';
-import { createEstimateFeeLimiter, setupAppMiddleware } from './expressapp/setup';
-import { registerTransactionRoutes } from './expressapp/transactions';
-import { registerWalletDataRoutes } from './expressapp/wallet-data';
-import { registerWalletRoutes } from './expressapp/wallets';
 import { logger } from './logger';
 import { AaveRouter } from './routes/aave';
+import { registerAdvertisementRoutes } from './routes/advertisements';
+import { registerBlockchainRoutes } from './routes/blockchain';
+import { createRouteHelpers } from './routes/context';
 import { error } from './routes/helpers';
+import { registerMaintenanceRoutes } from './routes/maintenance';
 import { createWalletLimiter } from './routes/middleware/createWalletLimiter';
+import { registerMoralisRoutes } from './routes/moralis';
+import { registerNotificationRoutes } from './routes/notifications';
+import { registerServiceRoutes } from './routes/services';
+import { createEstimateFeeLimiter, setupAppMiddleware } from './routes/setup';
+import { registerTransactionRoutes } from './routes/transactions';
 import { TssRouter } from './routes/tss';
+import { registerWalletDataRoutes } from './routes/walletdata';
+import { registerWalletRoutes } from './routes/wallets';
 import { WalletService } from './server';
 
 export class ExpressApp {
