+++ packages/bitcore-wallet-service/src/lib/routes/walletdata.ts
